pyproject.toml
[project]
name = "vps-scraper"
version = "0.2.0"
description = "VPS 测评文章爬虫 - 爬取 VPS 测评网站并使用 AI 提取结构化数据"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.12.0",
    "openai>=2.14.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "sniffio>=1.3.1",
    "zai-sdk>=0.2.0",
]

[project.scripts]
vps-scraper = "main:main"
